mcp-requirements.txt
# MCP Server Requirements - Fixed dependency conflicts

# Server - Updated for compatibility
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.7.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6

# Authentication
pyjwt>=2.8.0

# Networking
websockets>=12.0
aiohttp>=3.9.0
requests>=2.31.0

# Audio processing - essential only
gtts>=2.4.0
numpy>=1.24.0
scipy>=1.11.0

# Configuration
python-dotenv>=1.0.0

# Math support for calculator tool
sympy>=1.12

# Database connectivity (core)
pandas>=2.0.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
aiomysql>=0.2.0
aiosqlite>=0.19.0

# Optional: Advanced audio processing
# librosa>=0.10.0
# soundfile>=0.12.0
# pydub>=0.25.1

# Optional: Machine Learning (for local models)
# torch>=2.1.0
# torchaudio>=2.1.0
# transformers>=4.35.0
# accelerate>=0.24.0
# sentencepiece>=0.1.99

# Optional: Advanced TTS
# TTS>=0.21.0

# Optional: Snowflake support
# snowflake-connector-python>=3.6.0

# Utilities
tqdm>=4.66.0
packaging>=23.2
